refactor: package components -> attributes
