+++ src/main/java/com/festp/parsing/CommandParser.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import com.festp.styledmessage.components.Command;
-import com.festp.styledmessage.components.StyleAttribute;
+import com.festp.styledmessage.attributes.Command;
+import com.festp.styledmessage.attributes.StyleAttribute;
 import com.festp.utils.CommandValidator;
 import com.google.common.collect.Lists;
