+++ src/main/java/com/festp/styledmessage/attributes/CopyableText.java
@@ -1,4 +1,4 @@
-package com.festp.styledmessage.components;
+package com.festp.styledmessage.attributes;
 
 
 public class CopyableText implements StyleAttribute
