+++ src/main/java/com/festp/styledmessage/attributes/Formatting.java
@@ -1,4 +1,4 @@
-package com.festp.styledmessage.components;
+package com.festp.styledmessage.attributes;
 
 import org.bukkit.ChatColor;
