+++ src/main/java/com/festp/styledmessage/attributes/Link.java
@@ -1,4 +1,4 @@
-package com.festp.styledmessage.components;
+package com.festp.styledmessage.attributes;
 
 public class Link implements StyleAttribute
 {
