+++ src/main/java/com/festp/styledmessage/attributes/StyleAttribute.java
@@ -1,0 +1,4 @@
+package com.festp.styledmessage.attributes;
+
+public interface StyleAttribute {
+}
