--- src/main/java/com/festp/styledmessage/components/StyleAttribute.java
@@ -1,4 +1,0 @@
-package com.festp.styledmessage.components;
-
-public interface StyleAttribute {
-}
